:truck: Rework fixtures to be in their own files
